keep qt out of vc_core via PointCollections
vc_core and vc_tracer pulled in Qt only because VCCollection (a QObject)
carried the point data they read. split the data + IO into a Qt-free
PointCollections in vc_core; VCCollection is now a thin QObject shim that
emits its signals from the base's virtual hooks. drop the vc_ui lib and
the dead UDataManipulateUtils.
