+++ volume-cartographer/apps/VC3D/CMakeLists.txt
@@ -3,7 +3,12 @@ set(CMAKE_AUTORCC ON)
 set(CHUNK_CACHE_SIZE_GB "10" CACHE STRING "size of the chunk cache in gigabytes")
 add_definitions(-DCHUNK_CACHE_SIZE_GB=${CHUNK_CACHE_SIZE_GB})
 
+# Qt shim + Qt image helpers formerly in vc_ui; wrap the shim header for MOC.
+qt6_wrap_cpp(vccollection_moc ${CMAKE_SOURCE_DIR}/core/include/vc/ui/VCCollection.hpp)
+
 set(srcs
+    ${CMAKE_SOURCE_DIR}/core/src/VCCollection.cpp
+    ${vccollection_moc}
     VCAppMain.cpp
     CState.cpp
     CState.hpp
@@ -206,7 +211,6 @@ target_link_libraries(VC3D
     # mimalloc first so its symbols win malloc/free resolution.
     $<$<BOOL:${mimalloc_FOUND}>:mimalloc>
     vc_core
-    vc_ui
     vc_tracer
     vc_flattening
     Ceres::ceres
