+++ volume-cartographer/core/CMakeLists.txt
@@ -28,6 +28,8 @@ add_library(vc_core
         src/HttpFetch.cpp
         src/PostProcess.cpp
         src/Geometry.cpp
+        src/PointCollections.cpp
+        src/surface_metrics.cpp
         src/Thinning.cpp
         src/Umbilicus.cpp
         src/render/ChunkCache.cpp
@@ -75,10 +77,6 @@ set_source_files_properties(src/Slicing.cpp PROPERTIES
     COMPILE_OPTIONS "-O3;-ffast-math"
 )
 
-qt6_wrap_cpp(metrics_moc include/vc/ui/VCCollection.hpp)
-add_library(vc_ui src/VCCollection.cpp src/UDataManipulateUtils.cpp src/surface_metrics.cpp ${metrics_moc})
-target_link_libraries(vc_ui PUBLIC vc_core Qt6::Core Qt6::Gui Qt6::Widgets)
-
 add_library(vc_flattening src/ABFFlattening.cpp)
 target_link_libraries(vc_flattening PUBLIC vc_core OpenABF)
 
@@ -95,5 +93,5 @@ add_library(vc_tracer
     src/growth_strategies/ComponentPruning.cpp
     src/growth_strategies/GrowthConfig.cpp
 )
-target_link_libraries(vc_tracer PUBLIC vc_core vc_ui vc_flattening Ceres::ceres Qt6::Core)
+target_link_libraries(vc_tracer PUBLIC vc_core vc_flattening Ceres::ceres)
 target_include_directories(vc_tracer PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../libs/edt)
